+++ Tests/State/StoppedStateSpecs.swift
@@ -50,7 +50,9 @@ final class StoppedStateSpecs: QuickSpec {
             self.mockPlayer = MockCustomPlayer(overrideCurrentItem: self.item)
             self.mockPlayer.seekCompletionHandlerReturn = true
             self.delegate = MockPlayerContextDelegate()
-            self.playerContext = ModernAVPlayerContext(player: self.mockPlayer, audioSessionType: MockAudioSession.self, plugins: [self.plugin])
+            self.playerContext = ModernAVPlayerContext(player: self.mockPlayer,
+                                                       audioSession: MockAudioSession(),
+                                                       plugins: [self.plugin])
             self.playerContext.delegate = self.delegate
             self.playerContext.currentMedia = self.media
             self.tested = StoppedState(context: self.playerContext)